Move claude per-session ingest into ingest.ts
The trimmed-down `commands/claude.ts` no longer fronted a `burn claude`
command — it only held the single-session ingest helper called by the
claude harness adapter's `afterExit`. Keeping it under `commands/` was
misleading.

Move the helper to `ingest.ts` and rename it `ingestClaudeSession` so it
sits next to `ingestCodexSessions` / `ingestOpencodeSessions` as the
per-harness ingest peer. The claude harness adapter now imports from
`../ingest.js` like the other two adapters. Test moves to
`ingest-claude-session.test.ts` next to `ingest.ts`.
